+++ NEWS.md
@@ -14,6 +14,8 @@ When the target column was a factor where the levels were not in alphabetic orde
 
 * Fixes: Bug in `ROC` direction parameter being set wrong when `positive` is numeric. In regression tests, the `AUC` scores were *not* impacted.
 
+* In multinomial evaluation, the `Class Level Results` are sorted by the `Class`.
+
 * Imports `broom.mixed` to allow tidying of coefficients from `lme4::lmer` models.
 
 # cvms 1.0.2